// apps/web/src/app/core/models/ApplicationGroupUsersModel.ts
export type ApplicationGroupUsersCreateInput = {
  applicationGroupUserId: string;
  applicationGroupId: string;
  userId: string;
  applicationId: string;
  status: ApplicationGroupUserStatus;
  createdAt: Date;
  updatedAt: Date;
};
// ...
export interface IApplicationGroupUsers {
  applicationGroupUserId: string;
  applicationGroupId: string;
  userId: string;
  applicationId: string;
  status: ApplicationGroupUserStatus;
  createdAt: Date;
  updatedAt: Date;
}

export class ApplicationGroupUsers implements IApplicationGroupUsers {
  applicationGroupUserId: string;
  applicationGroupId: string;
  userId: string;
  applicationId: string;
  status: ApplicationGroupUserStatus;
  createdAt: Date;
  updatedAt: Date;

  constructor(data: Partial<IApplicationGroupUsers> = {}) {
    this.applicationGroupUserId = data.applicationGroupUserId ?? '';
    this.applicationGroupId = data.applicationGroupId ?? '';
    this.userId = data.userId ?? '';
    this.applicationId = data.applicationId ?? '';
    this.status = data.status ?? ApplicationGroupUserStatus.Active;
    this.createdAt = data.createdAt ?? new Date();
    this.updatedAt = data.updatedAt ?? new Date();
  }
}